Automatically update displayed pricing of adding when differing parent product cycles selected.
For some reason WHMCS doesn’t do this for us. When you change to an
annual cycle for a product, the attached addons change their price in
the cart, but not on the page, making it confusing what you’re paying.
This code updates the price on page to the correct amount based on the
selected cycle